Add files via upload
Individual parts of segments of data extracted from the .obj file into the excel file pasted into this word document (sorry that it is so convoluted of a path). It was the fastest way I could think of to insert the comma values into the data, since the replace function on Music Algorithms did not seem to work.